NFC - Expose hoistStaticallyBoundAllocationsInFunc
