+++ compiler/src/iree/compiler/Codegen/Transforms/Transforms.h
@@ -14,7 +14,6 @@
 
 #include "iree/compiler/Dialect/Flow/IR/FlowOps.h"
 #include "iree/compiler/Dialect/HAL/IR/HALOps.h"
-#include "llvm/Support/Debug.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
 #include "mlir/Dialect/Linalg/Transforms/Transforms.h"
 #include "mlir/Dialect/Vector/IR/VectorOps.h"
@@ -45,7 +44,7 @@ SliceAndDynamicDims cloneOffsetsSizesAndStrides(
 /// allocations creates an allocation, and inserts a subview to match the
 /// dynamic shape of the allocation. Returns std::nullopt if the method
 /// couldnt creat an allocation in the entry block.
-std::optional<Value> hoistStaticallyBoundAllocations(
+std::optional<Value> hoistOneStaticallyBoundAllocation(
     func::FuncOp funcOp, OpBuilder &builder, Location loc,
     MemRefType allocaType, ValueRange dynamicSizes,
     std::optional<uint64_t> alignment);
@@ -56,9 +55,11 @@ std::optional<Value> hoistStaticallyBoundAllocations(
 /// allocations creates an allocation, and inserts a subview to match the
 /// dynamic shape of the allocation. The method returns a value, but
 /// does not replace the uses of the `allocaOp`.
-std::optional<Value> hoistStaticallyBoundAllocations(func::FuncOp funcOp,
-                                                     OpBuilder &builder,
-                                                     memref::AllocaOp allocaOp);
+std::optional<Value> hoistOneStaticallyBoundAllocation(
+    func::FuncOp funcOp, OpBuilder &builder, memref::AllocaOp allocaOp);
+
+void hoistStaticallyBoundAllocationsInFunc(RewriterBase &rewriter,
+                                           func::FuncOp funcOp);
 
 /// Insert patterns to perform folding of AffineMinOp by matching the
 /// pattern generated by tile and distribute. Try to fold a affine.min op by
